Crafting System In Progress
There is a horizontal row of 5 selectable recipes at the top, which will hopefully scroll left and right to see more recipes. On the bottom left is the ingredients info (currently displayed through a list of text), and on the right will be buttons for normal crafting and bulk crafting, when implemented of course.

Right now, the menu needs to be toggled with input controls, the recipes row should be scrollable, the selectable item buttons should have text under them, and the ingredients info display should be more visual and use colors to indicate whether a particular item requirement is fulfilled or not. Finally, the crafting buttons have to be implemented, and there should be keyboard shortcuts for crafting without using the buttons.